Rename fenicsadapter to fenicsprecice. (#93)
* Follow naming convention discussed in #85.
* Rename package fenicsadapter to fenicsprecice.
* Use name FEniCS-preCICE adapter in setup.py description and README.md.
