--- fenicsadapter/__init__.py
@@ -1,2 +1,0 @@
-from .fenicsadapter import Adapter
-
