+++ fenicsprecice/__init__.py
@@ -1,0 +1,2 @@
+from .fenicsprecice import Adapter
+
