+++ tests/test_fenicsprecice.py
@@ -61,7 +61,7 @@ class TestCheckpointing(TestCase):
         """
         Test correct checkpoint storing
         """
-        import fenicsadapter
+        import fenicsprecice
         from precice import Interface, action_write_iteration_checkpoint
 
         def is_action_required_behavior(py_action):
@@ -79,7 +79,7 @@ class TestCheckpointing(TestCase):
         Interface.is_time_window_complete = MagicMock(return_value=True)
         Interface.advance = MagicMock()
 
-        precice = fenicsadapter.Adapter(self.dummy_config)
+        precice = fenicsprecice.Adapter(self.dummy_config)
 
         precice.store_checkpoint(self.u_n_mocked, self.t, self.n)
 
@@ -130,7 +130,7 @@ class TestExpressionHandling(TestCase):
         same points on a FEniCS Expression created by the Adapter
         """
         from precice import Interface
-        import fenicsadapter
+        import fenicsprecice
 
         Interface.get_dimensions = MagicMock(return_value=2)
         Interface.set_mesh_vertices = MagicMock(return_value=self.vertex_ids)
@@ -141,7 +141,7 @@ class TestExpressionHandling(TestCase):
 
         right_boundary = self.Right()
 
-        precice = fenicsadapter.Adapter(self.dummy_config)
+        precice = fenicsprecice.Adapter(self.dummy_config)
         precice._interface = Interface(None, None, None, None)
         precice.initialize(right_boundary, self.mesh, self.scalar_V)
         values = np.array([self.scalar_function(x, y) for x, y in zip(self.vertices_x, self.vertices_y)])
@@ -161,7 +161,7 @@ class TestExpressionHandling(TestCase):
         same points on a FEniCS Expression created by the Adapter
         """
         from precice import Interface
-        import fenicsadapter
+        import fenicsprecice
 
         Interface.get_dimensions = MagicMock(return_value=2)
         Interface.set_mesh_vertices = MagicMock(return_value=self.vertex_ids)
@@ -172,7 +172,7 @@ class TestExpressionHandling(TestCase):
 
         right_boundary = self.Right()
 
-        precice = fenicsadapter.Adapter(self.dummy_config)
+        precice = fenicsprecice.Adapter(self.dummy_config)
         precice._interface = Interface(None, None, None, None)
         precice.initialize(right_boundary, self.mesh, self.vector_V)
         values = np.array([self.vector_function(x, y) for x, y in zip(self.vertices_x, self.vertices_y)])
@@ -223,8 +223,8 @@ class TestExpressionHandling(TestCase):
 #         -------
 #         """
 #         from precice import Interface
-#         import fenicsadapter
-#         from fenicsadapter.adapter_core import FunctionType
+#         import fenicsprecice
+#         from fenicsprecice.adapter_core import FunctionType
 #
 #         Interface.get_dimensions = MagicMock(return_value=2)
 #         Interface.set_mesh_vertices = MagicMock(return_value=self.vertex_ids)
@@ -247,7 +247,7 @@ class TestExpressionHandling(TestCase):
 #             f_x_manual[key] = PointSource(self.V.sub(0), Point(px, py), dummy_forces_array[i, 0])
 #             f_y_manual[key] = PointSource(self.V.sub(1), Point(px, py), dummy_forces_array[i, 1])
 #
-#         precice = fenicsadapter.Adapter(self.dummy_config)
+#         precice = fenicsprecice.Adapter(self.dummy_config)
 #         precice._function_space = self.V
 #         precice._Dirichlet_Boundary = AutoSubDomain(clamped_boundary)
 #         precice._read_function_type = FunctionType.VECTOR
